refactor(gaussdb): move gaussdb resources to services
