+++ huaweicloud/provider.go
@@ -289,17 +289,17 @@ func Provider() *schema.Provider {
 			"huaweicloud_elb_flavors":                          dataSourceElbFlavorsV3(),
 			"huaweicloud_enterprise_project":                   DataSourceEnterpriseProject(),
 			"huaweicloud_fgs_dependencies":                     fgs.DataSourceFunctionGraphDependencies(),
-			"huaweicloud_gaussdb_cassandra_dedicated_resource": dataSourceGeminiDBDehResource(),
-			"huaweicloud_gaussdb_cassandra_instance":           dataSourceGeminiDBInstance(),
-			"huaweicloud_gaussdb_cassandra_instances":          dataSourceGeminiDBInstances(),
-			"huaweicloud_gaussdb_opengauss_instance":           dataSourceOpenGaussInstance(),
+			"huaweicloud_gaussdb_cassandra_dedicated_resource": gaussdb.DataSourceGeminiDBDehResource(),
+			"huaweicloud_gaussdb_cassandra_instance":           gaussdb.DataSourceGeminiDBInstance(),
+			"huaweicloud_gaussdb_cassandra_instances":          gaussdb.DataSourceGeminiDBInstances(),
+			"huaweicloud_gaussdb_opengauss_instance":           gaussdb.DataSourceOpenGaussInstance(),
 			"huaweicloud_gaussdb_opengauss_instances":          gaussdb.DataSourceOpenGaussInstances(),
-			"huaweicloud_gaussdb_mysql_configuration":          dataSourceGaussdbMysqlConfigurations(),
-			"huaweicloud_gaussdb_mysql_dedicated_resource":     dataSourceGaussDBMysqlDehResource(),
-			"huaweicloud_gaussdb_mysql_flavors":                dataSourceGaussdbMysqlFlavors(),
-			"huaweicloud_gaussdb_mysql_instance":               dataSourceGaussDBMysqlInstance(),
-			"huaweicloud_gaussdb_mysql_instances":              dataSourceGaussDBMysqlInstances(),
-			"huaweicloud_gaussdb_redis_instance":               dataSourceGaussRedisInstance(),
+			"huaweicloud_gaussdb_mysql_configuration":          gaussdb.DataSourceGaussdbMysqlConfigurations(),
+			"huaweicloud_gaussdb_mysql_dedicated_resource":     gaussdb.DataSourceGaussDBMysqlDehResource(),
+			"huaweicloud_gaussdb_mysql_flavors":                gaussdb.DataSourceGaussdbMysqlFlavors(),
+			"huaweicloud_gaussdb_mysql_instance":               gaussdb.DataSourceGaussDBMysqlInstance(),
+			"huaweicloud_gaussdb_mysql_instances":              gaussdb.DataSourceGaussDBMysqlInstances(),
+			"huaweicloud_gaussdb_redis_instance":               gaussdb.DataSourceGaussRedisInstance(),
 			"huaweicloud_identity_role":                        DataSourceIdentityRoleV3(),
 			"huaweicloud_identity_custom_role":                 DataSourceIdentityCustomRole(),
 			"huaweicloud_identity_group":                       iam.DataSourceIdentityGroup(),
@@ -456,10 +456,10 @@ func Provider() *schema.Provider {
 			"huaweicloud_evs_snapshot":                    ResourceEvsSnapshotV2(),
 			"huaweicloud_evs_volume":                      ResourceEvsStorageVolumeV3(),
 			"huaweicloud_fgs_function":                    fgs.ResourceFgsFunctionV2(),
-			"huaweicloud_gaussdb_cassandra_instance":      resourceGeminiDBInstanceV3(),
-			"huaweicloud_gaussdb_mysql_instance":          resourceGaussDBInstance(),
-			"huaweicloud_gaussdb_opengauss_instance":      resourceOpenGaussInstance(),
-			"huaweicloud_gaussdb_redis_instance":          resourceGaussRedisInstanceV3(),
+			"huaweicloud_gaussdb_cassandra_instance":      gaussdb.ResourceGeminiDBInstanceV3(),
+			"huaweicloud_gaussdb_mysql_instance":          gaussdb.ResourceGaussDBInstance(),
+			"huaweicloud_gaussdb_opengauss_instance":      gaussdb.ResourceOpenGaussInstance(),
+			"huaweicloud_gaussdb_redis_instance":          gaussdb.ResourceGaussRedisInstanceV3(),
 			"huaweicloud_ges_graph":                       ResourceGesGraphV1(),
 			"huaweicloud_identity_access_key":             resourceIdentityKey(),
 			"huaweicloud_identity_acl":                    ResourceIdentityACL(),
