+++ huaweicloud/services/acceptance/gaussdb/data_source_huaweicloud_gaussdb_cassandra_dedicated_resource_test.go
@@ -1,8 +1,9 @@
-package huaweicloud
+package gaussdb
 
 import (
 	"testing"
 
+	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/acceptance"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/utils/fmtp"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
@@ -11,8 +12,8 @@ import (
 
 func TestAccGeminiDBDehResourceDataSource_basic(t *testing.T) {
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:  func() { acceptance.TestAccPreCheck(t) },
+		Providers: acceptance.TestAccProviders,
 		Steps: []resource.TestStep{
 			{
 				Config: testAccGeminiDBDehResourceDataSource_basic,
