+++ huaweicloud/services/acceptance/gaussdb/data_source_huaweicloud_gaussdb_mysql_configuration_test.go
@@ -1,15 +1,17 @@
-package huaweicloud
+package gaussdb
 
 import (
 	"testing"
+
+	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/acceptance"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 )
 
 func TestAccGaussdbMysqlConfigurationDataSource_basic(t *testing.T) {
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:  func() { acceptance.TestAccPreCheck(t) },
+		Providers: acceptance.TestAccProviders,
 		Steps: []resource.TestStep{
 			{
 				Config: testAccGaussdbMysqlConfigurationDataSource_basic,
