+++ huaweicloud/services/acceptance/gaussdb/resource_huaweicloud_gaussdb_mysql_instance_test.go
@@ -1,9 +1,10 @@
-package huaweicloud
+package gaussdb
 
 import (
 	"fmt"
 	"testing"
 
+	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/acceptance"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/utils/fmtp"
 
 	"github.com/chnsz/golangsdk/openstack/taurusdb/v3/instances"
@@ -20,8 +21,8 @@ func TestAccGaussDBInstance_basic(t *testing.T) {
 	resourceName := "huaweicloud_gaussdb_mysql_instance.test"
 
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
+		PreCheck:     func() { acceptance.TestAccPreCheck(t) },
+		Providers:    acceptance.TestAccProviders,
 		CheckDestroy: testAccCheckGaussDBInstanceDestroy,
 		Steps: []resource.TestStep{
 			{
@@ -36,8 +37,8 @@ func TestAccGaussDBInstance_basic(t *testing.T) {
 }
 
 func testAccCheckGaussDBInstanceDestroy(s *terraform.State) error {
-	config := testAccProvider.Meta().(*config.Config)
-	client, err := config.GaussdbV3Client(HW_REGION_NAME)
+	config := acceptance.TestAccProvider.Meta().(*config.Config)
+	client, err := config.GaussdbV3Client(acceptance.HW_REGION_NAME)
 	if err != nil {
 		return fmtp.Errorf("Error creating HuaweiCloud GaussDB client: %s", err)
 	}
@@ -67,8 +68,8 @@ func testAccCheckGaussDBInstanceExists(n string, instance *instances.TaurusDBIns
 			return fmtp.Errorf("No ID is set.")
 		}
 
-		config := testAccProvider.Meta().(*config.Config)
-		client, err := config.GaussdbV3Client(HW_REGION_NAME)
+		config := acceptance.TestAccProvider.Meta().(*config.Config)
+		client, err := config.GaussdbV3Client(acceptance.HW_REGION_NAME)
 		if err != nil {
 			return fmtp.Errorf("Error creating HuaweiCloud GaussDB client: %s", err)
 		}
