+++ huaweicloud/services/acceptance/gaussdb/resource_huaweicloud_gaussdb_redis_instance_test.go
@@ -1,9 +1,10 @@
-package huaweicloud
+package gaussdb
 
 import (
 	"fmt"
 	"testing"
 
+	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/acceptance"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/utils/fmtp"
 
 	"github.com/chnsz/golangsdk/openstack/geminidb/v3/instances"
@@ -22,8 +23,8 @@ func TestAccGaussRedisInstance_basic(t *testing.T) {
 	resourceName := "huaweicloud_gaussdb_redis_instance.test"
 
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
+		PreCheck:     func() { acceptance.TestAccPreCheck(t) },
+		Providers:    acceptance.TestAccProviders,
 		CheckDestroy: testAccCheckGaussRedisInstanceDestroy,
 		Steps: []resource.TestStep{
 			{
@@ -53,8 +54,8 @@ func TestAccGaussRedisInstance_basic(t *testing.T) {
 }
 
 func testAccCheckGaussRedisInstanceDestroy(s *terraform.State) error {
-	config := testAccProvider.Meta().(*config.Config)
-	client, err := config.GeminiDBV3Client(HW_REGION_NAME)
+	config := acceptance.TestAccProvider.Meta().(*config.Config)
+	client, err := config.GeminiDBV3Client(acceptance.HW_REGION_NAME)
 	if err != nil {
 		return fmtp.Errorf("Error creating HuaweiCloud GaussRedis client: %s", err)
 	}
@@ -87,8 +88,8 @@ func testAccCheckGaussRedisInstanceExists(n string, instance *instances.GeminiDB
 			return fmtp.Errorf("No ID is set.")
 		}
 
-		config := testAccProvider.Meta().(*config.Config)
-		client, err := config.GeminiDBV3Client(HW_REGION_NAME)
+		config := acceptance.TestAccProvider.Meta().(*config.Config)
+		client, err := config.GeminiDBV3Client(acceptance.HW_REGION_NAME)
 		if err != nil {
 			return fmtp.Errorf("Error creating HuaweiCloud GuassRedis client: %s", err)
 		}
