+++ huaweicloud/services/gaussdb/resource_huaweicloud_gaussdb_cassandra_instance.go
@@ -1,4 +1,4 @@
-package huaweicloud
+package gaussdb
 
 import (
 	"fmt"
@@ -14,13 +14,14 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
+	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/common"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/config"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/utils"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/utils/fmtp"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/utils/logp"
 )
 
-func resourceGeminiDBInstanceV3() *schema.Resource {
+func ResourceGeminiDBInstanceV3() *schema.Resource {
 	return &schema.Resource{
 		Create: resourceGeminiDBInstanceV3Create,
 		Read:   resourceGeminiDBInstanceV3Read,
@@ -247,7 +248,7 @@ func resourceGeminiDBInstanceV3() *schema.Resource {
 				}, false),
 			},
 
-			"tags": tagsSchema(),
+			"tags": common.TagsSchema(),
 		},
 	}
 }
@@ -312,13 +313,13 @@ func GeminiDBInstanceStateRefreshFunc(client *golangsdk.ServiceClient, instanceI
 
 func resourceGeminiDBInstanceV3Create(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*config.Config)
-	client, err := config.GeminiDBV3Client(GetRegion(d, config))
+	client, err := config.GeminiDBV3Client(config.GetRegion(d))
 	if err != nil {
 		return fmtp.Errorf("Error creating HuaweiCloud GeminiDB client: %s ", err)
 	}
 
 	// If force_import set, try to import it instead of creating
-	if hasFilledOpt(d, "force_import") {
+	if _, ok := d.GetOk("force_import"); ok {
 		logp.Printf("[DEBUG] Gaussdb cassandra instance force_import is set, try to import it instead of creating")
 		listOpts := instances.ListGeminiDBInstanceOpts{
 			Name: d.Get("name").(string),
@@ -342,13 +343,13 @@ func resourceGeminiDBInstanceV3Create(d *schema.ResourceData, meta interface{}) 
 
 	createOpts := instances.CreateGeminiDBOpts{
 		Name:                d.Get("name").(string),
-		Region:              GetRegion(d, config),
+		Region:              config.GetRegion(d),
 		AvailabilityZone:    d.Get("availability_zone").(string),
 		VpcId:               d.Get("vpc_id").(string),
 		SubnetId:            d.Get("subnet_id").(string),
 		SecurityGroupId:     d.Get("security_group_id").(string),
 		ConfigurationId:     d.Get("configuration_id").(string),
-		EnterpriseProjectId: GetEnterpriseProjectID(d, config),
+		EnterpriseProjectId: config.GetEnterpriseProjectID(d),
 		DedicatedResourceId: d.Get("dedicated_resource_id").(string),
 		Mode:                "Cluster",
 		Flavor:              resourceGeminiDBFlavor(d),
@@ -361,7 +362,7 @@ func resourceGeminiDBInstanceV3Create(d *schema.ResourceData, meta interface{}) 
 
 	// PrePaid
 	if d.Get("charging_mode") == "prePaid" {
-		if err := validatePrePaidChargeInfo(d); err != nil {
+		if err := common.ValidatePrePaidChargeInfo(d); err != nil {
 			return err
 		}
 
@@ -418,7 +419,7 @@ func resourceGeminiDBInstanceV3Create(d *schema.ResourceData, meta interface{}) 
 
 func resourceGeminiDBInstanceV3Read(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*config.Config)
-	client, err := config.GeminiDBV3Client(GetRegion(d, config))
+	client, err := config.GeminiDBV3Client(config.GetRegion(d))
 	if err != nil {
 		return fmtp.Errorf("Error creating HuaweiCloud GeminiDB client: %s", err)
 	}
@@ -426,7 +427,7 @@ func resourceGeminiDBInstanceV3Read(d *schema.ResourceData, meta interface{}) er
 	instanceID := d.Id()
 	instance, err := instances.GetInstanceByID(client, instanceID)
 	if err != nil {
-		return CheckDeleted(d, err, "GeminiDB")
+		return common.CheckDeleted(d, err, "GeminiDB")
 	}
 	if instance.Id == "" {
 		d.SetId("")
@@ -518,14 +519,14 @@ func resourceGeminiDBInstanceV3Read(d *schema.ResourceData, meta interface{}) er
 
 func resourceGeminiDBInstanceV3Delete(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*config.Config)
-	client, err := config.GeminiDBV3Client(GetRegion(d, config))
+	client, err := config.GeminiDBV3Client(config.GetRegion(d))
 	if err != nil {
 		return fmtp.Errorf("Error creating HuaweiCloud GeminiDB client: %s ", err)
 	}
 
 	instanceId := d.Id()
 	if d.Get("charging_mode") == "prePaid" {
-		if err := UnsubscribePrePaidResource(d, config, []string{instanceId}); err != nil {
+		if err := common.UnsubscribePrePaidResource(d, config, []string{instanceId}); err != nil {
 			// Try to delete resource directly when unsubscrbing failed
 			res := instances.Delete(client, instanceId)
 			if res.Err != nil {
@@ -561,11 +562,11 @@ func resourceGeminiDBInstanceV3Delete(d *schema.ResourceData, meta interface{}) 
 
 func resourceGeminiDBInstanceV3Update(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*config.Config)
-	client, err := config.GeminiDBV3Client(GetRegion(d, config))
+	client, err := config.GeminiDBV3Client(config.GetRegion(d))
 	if err != nil {
 		return fmtp.Errorf("Error creating Huaweicloud Vpc: %s", err)
 	}
-	bssClient, err := config.BssV2Client(GetRegion(d, config))
+	bssClient, err := config.BssV2Client(config.GetRegion(d))
 	if err != nil {
 		return fmtp.Errorf("Error creating HuaweiCloud bss V2 client: %s", err)
 	}
